pyproject.toml
[tool.poetry]
name = "RSS to Jukebox Extended"
version = "1.0.0"
description = "Generates a Minecraft resource pack and JEXT discs.json"
authors = ["Ralph <ralph@podaboutli.st>"]
readme = "README.md"
license = "AGPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "1.0.0"
requests = "^2.31.0"
rss-parser = "^1.1.1"
python-ffmpeg = "^2.0.4"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
ruff = "^0.0.292"
pre-commit = "^3.5.0"
mypy = "^1.6.0"
types-requests = "^2.31.0.9"


[tool.poetry.group.optional.dependencies]
poethepoet = "^0.24.1"

[tool.ruff]
select = ["B", "D", "E", "F", "I", "N", "S", "W", "UP", "TCH"]
ignore = ["D103", "D211", "D213", "D400"]
line-length = 88
target-version = "py311"
src = ["src"]

[tool.ruff.format]
line-ending = "lf"
